MSCSendToOwnersConsensusComparisonSpec: #after >= #before
In this test an amount of 100.0 MSC is sent to all MSC owners and it is
confirmed that the balance of the recipients really increased. However,
not every MSC owner may necessarily get a piece of the pie in this test,
and this is especially the case for owners with very small amounts. For
those there is no balance increase, which lead to test failures in some
cases.