Set transaction fee and assert transaction relay fees in regtest mode
While there is a way to set the transaction fee (as done), there is no way to set the relay fee, but some tests use that value, and more are likely using it in the future.

The main reason, why the relay fee is not simply extracted from `getinfo` and used directly (which I tried):

There is actually a small difference in the dust limit calculation between Bitcoin Core 0.9 and Bitcoin Core 0.10, resulting in slightly different limits due to rounding effects. This can cause unexpected results, when using some specific settings, depending on which client is used.

Until it is properly addressed (in omnicore-0.0.10), this route ensures no unexpected situation can be triggered.