Add specification for the grace period of feature activations
Features are activated at specific block heights, which must be within
the range of a grace period to ensure users have enough time to update
their clients.

The use of activation commands is restricted to whitelisted senders.

To whitelist a source to allow feature activations, the option
-omniactivationallowsender="sender" can be used.

The feature identifier 3 is currently unused, and a good candidate for
tests.

Note: this test is only successful with a clean state!