Add specification for the deactivation of DEx "over-offers"
Once the feature with identifier 4 is active, it is no longer allowed
to create offers on the traditional distributed exchange, for an amount
that is more than the available balance.

During the grace period, the old rules are still in place.

Note: this test is only successful with a clean state!