Deactivate all feature activation related tests
Feature activation tests currently affect other tests and can only be
executed with a prestine state.

This doesn't play well with automated tests, and may not be obvious,
when using OmniJ to run tests via an IDE.

The "MetaDExActivationSpec" only works, if the MetaDEx is not yet
activated, which is not the default in Omni Core, even in regtest mode,
so it's required to manually un-ignore the specification.