Add workaround to avoid conflict of activations (zathras-crypto)
To avoid that the activations affect other tests, a workaround is used
to revert consensus affecting changes:

Before the tests, the hash of a "marker block" is stored, and after the
tests 50 blocks are mined, and finally the "marker block" is
invalidated. This results in a reorganization, and because only the
state of the last 50 blocks is persisted, it completely resets the
state, including all activations inbetween.

This allows to enable 2/3 activation specs.