Add new test to confirm STODB is reorg-safe
To confirm that the database for "send to owners" transactions is
reorg-safe, two owners and an actor are created. The actor sends two
STO transactions, whereby the second one is invalidated and reverted.
When then a third STO transaction is sent, it is expected that the
information for the first (and third) transaction is correct.