Add improved base class for integration tests and enforce order
The new class "IntegrationTest" is used to set reusable properties,
such as the test log formatting.

Test failures are described in greater detail, in a similar fashion as
in the test reports, where the actual cause (e.g. is failed condition)
is printed.

The new class "RegTest" is intended to be used for RPC tests in regtest
mode.

Feature activation tests are split into a seperate task, which is
executed before running the ":omnij-rpc:regTest" task. This ensures
features, if necessary, are activated before the tests, which may
depend on the feature, are executed.