Add activation tests for all-pair trading on MetaDEx
Adds test for the activation of all-pair trading on the distributed token exchange.

Once the feature with identifier 8 is active, the distributed token exchange can be used with non-Omni tokens.

During the grace period, the old rules are still in place.

Note: this test is only successful with a clean state, and requires that the feature is initially disabled!