Fix error in expected value in LargestSlice test
The data was wrong but the error wasn’t detected until the recent change
made to addressVerifyInfo a few commits ago. The JSON test data was using
`reserved` for reserved amounts and only with the recent change to 
AddressVerifyInfo to use `reserved` instead of `reserved_balance` did 
the problem appear. Note that 1,335,000,000 is the correct total of USDT.