Add TransactionInfo.transactionType(), JavaDoc comments and tests
TransactionInfo.transactionType() returns Optional<TransactionType>

Optional<TransactionType> is the recommended type for handling a
TransactionType enum. With the forthcoming "Pattern Matching for Switch" it will be possible to use it in switch statements very elegantly.

Tests have been added as well as extensive JavaDoc with code examples
for how to use this variable type in earlier (and admittedly more verbose) versions of Java.