WIP: OmniLiteSend
* Distinguish between Keychain sending/signing service and
  RpcClient sending/signing service.

Keychain services use a local HD keychain and a JSON-RPC connection
to an Omni server. The idea is the Omni server powers multiple wallets.

RpcClient services use a JSON-RPC connection to a Bitcoin server (and
should work with a full-pruned node as well as the ConsensusJ SPV daemon)
The idea is the JSON-RPC server powers a single wallet.
 
SendTool uses the KeyChain-based services

OmniLiteSend uses the RpcClient-based service and has been tested with
the ConsensusJ SPV daemon.